refactor: extract StrategiesConfiguratorOffchain
